move schedule.py to /tasks
