+++ redash/tasks/__init__.py
@@ -3,4 +3,5 @@ from .queries import (QueryTask, enqueue_query, execute_query, refresh_queries,
                       refresh_schemas, cleanup_query_results, empty_schedules)
 from .alerts import check_alerts_for_query
 from .failure_report import send_aggregated_errors
-from .hard_limiting_worker import HardLimitingWorker
+from .hard_limiting_worker import *
+from .schedule import *
